+++ tests/fixtures/modules/terraform-azure-avm-res-mock/examples/default/tflint-pre.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-echo "Running tflint-pre.sh for default example"